Task 2/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
torch
facenet-pytorch
python-multipart
scikit-learn
